fix: restructure the folder and html structure
