--- src/components/Body.jsx
@@ -1,15 +1,0 @@
-import Footer from "./Footer";
-import Header from "./Header";
-import Workouts from "./Workouts";
-
-const Body = () => {
-  return (
-    <div className="">
-      <Header />
-      <Workouts />
-      <Footer />
-    </div>
-  );
-};
-
-export default Body;
