+++ src/pages/Contact.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Contact() {
+  return <div>Contact</div>;
+}
+
+export default Contact;
