+++ src/pages/Diet.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Diet() {
+  return <div>Diet</div>;
+}
