+++ src/pages/Home.jsx
@@ -1,0 +1,15 @@
+import Footer from "../components/Footer";
+import Header from "../components/Header";
+import Workouts from "../components/Workouts";
+
+const Home = () => {
+  return (
+    <div className="">
+      <Header />
+      <Workouts />
+      <Footer />
+    </div>
+  );
+};
+
+export default Home;
