+++ src/pages/Workout.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Workout() {
+  return <div>Workout</div>;
+}
+
+export default Workout;
